+++ caliban/cluster/__init__.py
@@ -1,2 +1,2 @@
 from .cluster import (Cluster, default_credentials, credentials_from_file,
-                      credentials_from_env, parser, run_cli_command)
+                      parser, run_cli_command)